Extract Android cloud sign-in test support
